Index resources in the right folder + rename tests folder
