+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(AirQuality)
 
 set(CMAKE_CXX_STANDARD 11)
 #set(CMAKE_CXX_FLAGS "-DDEBUG -ansi -pedantic -Wall")
-add_executable(AirQuality src/model/User.cpp src/model/User.h src/model/Individual.cpp src/model/Individual.h src/model/Cleaner.cpp src/model/Cleaner.h src/model/Provider.cpp src/model/Provider.h src/model/Government.cpp src/model/Government.h src/model/Sensor.cpp src/model/Sensor.h src/model/Measure.cpp src/model/Measure.h src/model/Attribute.cpp src/model/Attribute.h src/interface/IndividualInterface.cpp src/interface/IndividualInterface.h src/interface/GovernmentInterface.cpp src/interface/GovernmentInterface.h src/interface/ProviderInterface.cpp src/interface/ProviderInterface.h src/services/Stats.cpp src/services/Stats.h src/services/Clustering.cpp src/services/Clustering.h src/services/ClearAnalysis.cpp src/services/ClearAnalysis.h src/services/Integrity.cpp src/services/Integrity.h src/data/Reader.cpp src/data/Reader.h Data/DataAccess.cpp src/data/DataAccess.h src/interface/InterfaceUtils.cpp src/interface/InterfaceUtils.h Tests/main.cpp Tests/IntegrityTest.cpp Tests/IntegrityTest.h)
+add_executable(AirQuality src/model/User.cpp src/model/User.h src/model/Individual.cpp src/model/Individual.h src/model/Cleaner.cpp src/model/Cleaner.h src/model/Provider.cpp src/model/Provider.h src/model/Government.cpp src/model/Government.h src/model/Sensor.cpp src/model/Sensor.h src/model/Measure.cpp src/model/Measure.h src/model/Attribute.cpp src/model/Attribute.h src/interface/IndividualInterface.cpp src/interface/IndividualInterface.h src/interface/GovernmentInterface.cpp src/interface/GovernmentInterface.h src/interface/ProviderInterface.cpp src/interface/ProviderInterface.h src/services/Stats.cpp src/services/Stats.h src/services/Clustering.cpp src/services/Clustering.h src/services/ClearAnalysis.cpp src/services/ClearAnalysis.h src/services/Integrity.cpp src/services/Integrity.h src/data/Reader.cpp src/data/Reader.h Data/DataAccess.cpp src/data/DataAccess.h src/interface/InterfaceUtils.cpp src/interface/InterfaceUtils.h tests/main.cpp tests/IntegrityTest.cpp tests/IntegrityTest.h)
