+++ resources/release/cleaners.csv
@@ -1,0 +1,2 @@
+Cleaner0;45.333333;1.333333;2019-02-01 12:00:00;2019-03-01 00:00:00;
+Cleaner1;46.666667;3.666667;2019-02-01 12:00:00;2019-03-01 00:00:00;
