+++ resources/release/providers.csv
@@ -1,0 +1,2 @@
+Provider0;Cleaner0;
+Provider1;Cleaner1;
