+++ resources/release/sensors.csv
@@ -1,0 +1,100 @@
+Sensor0;44;-1;
+Sensor1;44;-0.3;
+Sensor2;44;0.4;
+Sensor3;44;1.1;
+Sensor4;44;1.8;
+Sensor5;44;2.5;
+Sensor6;44;3.2;
+Sensor7;44;3.9;
+Sensor8;44;4.6;
+Sensor9;44;5.3;
+Sensor10;44.4;-1;
+Sensor11;44.4;-0.3;
+Sensor12;44.4;0.4;
+Sensor13;44.4;1.1;
+Sensor14;44.4;1.8;
+Sensor15;44.4;2.5;
+Sensor16;44.4;3.2;
+Sensor17;44.4;3.9;
+Sensor18;44.4;4.6;
+Sensor19;44.4;5.3;
+Sensor20;44.8;-1;
+Sensor21;44.8;-0.3;
+Sensor22;44.8;0.4;
+Sensor23;44.8;1.1;
+Sensor24;44.8;1.8;
+Sensor25;44.8;2.5;
+Sensor26;44.8;3.2;
+Sensor27;44.8;3.9;
+Sensor28;44.8;4.6;
+Sensor29;44.8;5.3;
+Sensor30;45.2;-1;
+Sensor31;45.2;-0.3;
+Sensor32;45.2;0.4;
+Sensor33;45.2;1.1;
+Sensor34;45.2;1.8;
+Sensor35;45.2;2.5;
+Sensor36;45.2;3.2;
+Sensor37;45.2;3.9;
+Sensor38;45.2;4.6;
+Sensor39;45.2;5.3;
+Sensor40;45.6;-1;
+Sensor41;45.6;-0.3;
+Sensor42;45.6;0.4;
+Sensor43;45.6;1.1;
+Sensor44;45.6;1.8;
+Sensor45;45.6;2.5;
+Sensor46;45.6;3.2;
+Sensor47;45.6;3.9;
+Sensor48;45.6;4.6;
+Sensor49;45.6;5.3;
+Sensor50;46.0;-1;
+Sensor51;46.0;-0.3;
+Sensor52;46.0;0.4;
+Sensor53;46.0;1.1;
+Sensor54;46.0;1.8;
+Sensor55;46.0;2.5;
+Sensor56;46.0;3.2;
+Sensor57;46.0;3.9;
+Sensor58;46.0;4.6;
+Sensor59;46.0;5.3;
+Sensor60;46.4;-1;
+Sensor61;46.4;-0.3;
+Sensor62;46.4;0.4;
+Sensor63;46.4;1.1;
+Sensor64;46.4;1.8;
+Sensor65;46.4;2.5;
+Sensor66;46.4;3.2;
+Sensor67;46.4;3.9;
+Sensor68;46.4;4.6;
+Sensor69;46.4;5.3;
+Sensor70;46.8;-1;
+Sensor71;46.8;-0.3;
+Sensor72;46.8;0.4;
+Sensor73;46.8;1.1;
+Sensor74;46.8;1.8;
+Sensor75;46.8;2.5;
+Sensor76;46.8;3.2;
+Sensor77;46.8;3.9;
+Sensor78;46.8;4.6;
+Sensor79;46.8;5.3;
+Sensor80;47.2;-1;
+Sensor81;47.2;-0.3;
+Sensor82;47.2;0.4;
+Sensor83;47.2;1.1;
+Sensor84;47.2;1.8;
+Sensor85;47.2;2.5;
+Sensor86;47.2;3.2;
+Sensor87;47.2;3.9;
+Sensor88;47.2;4.6;
+Sensor89;47.2;5.3;
+Sensor90;47.6;-1;
+Sensor91;47.6;-0.3;
+Sensor92;47.6;0.4;
+Sensor93;47.6;1.1;
+Sensor94;47.6;1.8;
+Sensor95;47.6;2.5;
+Sensor96;47.6;3.2;
+Sensor97;47.6;3.9;
+Sensor98;47.6;4.6;
+Sensor99;47.6;5.3;
