+++ resources/release/users.csv
@@ -1,0 +1,2 @@
+User0;Sensor70;
+User1;Sensor36;
