+++ resources/tests/sensors.csv
@@ -1,0 +1,14 @@
+Sensor0;44;-1;
+Sensor1;44;-0.3;
+Sensor2;44;0.4;
+Sensor3;44;1.1;
+Sensor4;44;1.8;
+Sensor5;44;2.5;
+Sensor6;44;3.2;
+Sensor7;44;3.9;
+Sensor8;44;4.6;
+Sensor9;44;5.3;
+Sensor10;0.0;0.0;
+Sensor11;10.0;20.0;
+Sensor12;44;2.0;
+Sensor13;44;2.2;
