+++ resources/tests/users.csv
@@ -1,0 +1,4 @@
+User0;Sensor70;
+User1;Sensor36;
+User2;Sensor12;
+User3;Sensor13;
