+++ tests/IntegrityTest.h
@@ -1,0 +1,16 @@
+//
+// Created by allan guigal on 31/05/2021.
+//
+
+#ifndef AIRQUALITY_INTEGRITYTEST_H
+#define AIRQUALITY_INTEGRITYTEST_H
+
+using namespace std;
+
+class IntegrityTest {
+public:
+    static void test();
+};
+
+
+#endif //AIRQUALITY_INTEGRITYTEST_H
